Extract database queries
