+++ src/database/mod.rs
@@ -1,7 +1,9 @@
-//! src/utils/database/mod.rs
+//! src/database/mod.rs
 
 pub mod connection;
+pub mod queries;
 pub mod schema;
 
 pub use connection::*;
+pub use queries::*;
 pub use schema::*;
